修改主键泛型为Int类型

+++ src/main/java/com/lord/advertis/dao/UsersDao.java
@@ -3,6 +3,6 @@ package com.lord.advertis.dao;
 import com.lord.advertis.base.BaseMapper;
 import com.lord.advertis.domain.Users;
 
-public interface UsersDao extends BaseMapper<Users, Long>{
+public interface UsersDao extends BaseMapper<Users, Integer>{
 
 }
